move solidity version selector to compile tab
